Add files via upload
Nernst Equation is used to determine the resting potentials of individual ions (Sodium, Potassium, Chloride,etc.). In  real cell, membrane voltage is contributed by many ions and the value of the membrane potential is determined by the concentration gradients and the relative permeability values of ions.  Relative Permeability is the easiness of an ion to cross the membrane.